+++ ruma-federation-api/src/device.rs
@@ -1,0 +1,2 @@
+//! Endpoints to retrieve information about user devices
+pub mod get_devices;