+++ ruma-federation-api/src/device/get_devices/mod.rs
@@ -1,0 +1,2 @@
+//! Endpoint to get information about a user's devices
+pub mod v1;